+++ bukkit/src/main/java/com/github/games647/fastlogin/bukkit/hooks/xAuthHook.java
@@ -2,9 +2,6 @@ package com.github.games647.fastlogin.bukkit.hooks;
 
 import de.luricos.bukkit.xAuth.xAuth;
 import de.luricos.bukkit.xAuth.xAuthPlayer;
-import de.luricos.bukkit.xAuth.xAuthPlayer.Status;
-
-import java.sql.Timestamp;
 
 import org.bukkit.entity.Player;
 
@@ -20,19 +17,12 @@ public class xAuthHook implements AuthPlugin {
 
         xAuthPlayer xAuthPlayer = xAuthPlugin.getPlayerManager().getPlayer(player);
         if (xAuthPlayer != null) {
+            xAuthPlugin.getPlayerManager().doLogin(xAuthPlayer);
             //we checked that the player is premium (paid account)
             xAuthPlayer.setPremium(true);
+
             //mark the player online
             xAuthPlugin.getAuthClass(xAuthPlayer).online(xAuthPlayer.getName());
-
-            //update last login time
-            xAuthPlayer.setLoginTime(new Timestamp(System.currentTimeMillis()));
-
-            //mark the player as logged in
-            xAuthPlayer.setStatus(Status.AUTHENTICATED);
-
-            //restore inventory
-            xAuthPlugin.getPlayerManager().unprotect(xAuthPlayer);
         }
     }
 }